+++ dcicutils/qa_utils.py
@@ -17,6 +17,7 @@ import uuid
 import warnings
 
 from json import dumps as json_dumps, loads as json_loads
+from unittest import mock
 from .misc_utils import PRINT, ignored, Retry, CustomizableProperty, getattr_customized, remove_prefix, REF_TZ
 
 
@@ -430,14 +431,39 @@ class MockFileWriter:
 class MockFileSystem:
     """Extremely low-tech mock file system."""
 
-    def __init__(self, files=None, default_encoding='utf-8'):
+    def __init__(self, files=None, default_encoding='utf-8', auto_mirror_files_for_read=False, do_not_auto_mirror=()):
         self.default_encoding = default_encoding
+        # Setting this dynamically will make things inconsistent
+        self._auto_mirror_files_for_read = auto_mirror_files_for_read
+        self._do_not_auto_mirror = set(do_not_auto_mirror or [])
         self.files = {filename: content.encode(default_encoding) for filename, content in (files or {}).items()}
+        for filename in self.files:
+            self._do_not_mirror(filename)
+
+    IO_OPEN = staticmethod(io.open)
+    OS_PATH_EXISTS = staticmethod(os.path.exists)
+    OS_REMOVE = staticmethod(os.remove)
+
+    def _do_not_mirror(self, file):
+        if self._auto_mirror_files_for_read:
+            self._do_not_auto_mirror.add(file)
+
+    def _maybe_auto_mirror_file(self, file):
+        if self._auto_mirror_files_for_read:
+            if file not in self._do_not_auto_mirror:
+                if (self.OS_PATH_EXISTS(file)
+                        # file might be in files if someone has been manipulating the file structure directly
+                        and file not in self.files):
+                    with open(file, 'rb') as fp:
+                        self.files[file] = fp.read()
+                self._do_not_mirror(file)
 
     def exists(self, file):
+        self._maybe_auto_mirror_file(file)
         return bool(self.files.get(file))
 
     def remove(self, file):
+        self._maybe_auto_mirror_file(file)
         if not self.files.pop(file, None):
             raise FileNotFoundError("No such file or directory: %s" % file)
 
@@ -456,6 +482,7 @@ class MockFileSystem:
             raise AssertionError("Mocked io.open doesn't handle mode=%r." % mode)
 
     def _open_for_read(self, file, binary=False, encoding=None):
+        self._do_not_mirror(file)
         content = self.files.get(file)
         if content is None:
             raise FileNotFoundError("No such file or directory: %s" % file)
@@ -464,8 +491,16 @@ class MockFileSystem:
         return io.BytesIO(content) if binary else io.StringIO(content.decode(encoding or self.default_encoding))
 
     def _open_for_write(self, file_system, file, binary=False, encoding=None):
+        self._do_not_mirror(file)
         return MockFileWriter(file_system=file_system, file=file, binary=binary,
                               encoding=encoding or self.default_encoding)
+
+    @contextlib.contextmanager
+    def mock_exists_open_remove(self):
+        with mock.patch("os.path.exists", self.exists):
+            with mock.patch("io.open", self.open):
+                with mock.patch("os.remove", self.remove):
+                    yield self
 
 
 class MockUUIDModule: